🚨 Fix progress callback lint findings
Move owned callback parameters into algorithms and layout writers. Add direct includes, remove unused headers, and use explicit initialization and const qualifiers.

Verify worker cleanup with an explicit exception assertion. All 10 affected C++ test targets and prek run -a pass.

+++ include/fiction/physical_design/exact.hpp
@@ -215,7 +215,7 @@ class exact_impl
             ps{std::move(p)},
             pst{st},
             progress{ps.on_progress, "aspect ratios"},
-            worker_progress{ps.on_worker_progress, std::max(std::size_t{1}, static_cast<std::size_t>(ps.num_threads))},
+            worker_progress{ps.on_worker_progress, std::max(std::size_t{1}, ps.num_threads)},
             scheme{std::move(clocking_scheme)},
             black_list{sbl}
     {
